Decouple bootstrap and post targets
Something weird is happening with the post target, where most of the
time, but not all of the time, the target fails with pacman errors about
packages missing signatures. Yet the steps before that worked without
any such issues, and when re-run, or run with different sets of packages
yields a different result. So for the time being I will remove the auto
call to post that was in the bootstrap target, and make this an optional
extra until I can get to the bottom of this.